Add unit tests and restructure tests and add support to inject read and write actions
