+++ Tests/BluetoothServicesTests/HealthThermometerTests.swift
@@ -1,0 +1,42 @@
+//
+// This source file is part of the Stanford Spezi open source project
+//
+// SPDX-FileCopyrightText: 2024 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+@_spi(TestingSupport)
+@testable import BluetoothServices
+import CoreBluetooth
+import NIO
+@_spi(TestingSupport)
+@testable import SpeziBluetooth
+import XCTByteCoding
+import XCTest
+
+
+final class HealthThermometerTests: XCTestCase {
+    func testMeasurementInterval() throws {
+        try testIdentity(from: MeasurementInterval.noPeriodicMeasurement)
+        try testIdentity(from: MeasurementInterval.duration(24))
+    }
+
+    func testTemperatureMeasurement() throws {
+        let data: UInt32 = 0xAFAFAFAF // 4 bytes for the medfloat
+        let time = DateTime(hours: 13, minutes: 12, seconds: 12)
+
+        try testIdentity(from: TemperatureMeasurement(temperature: data, unit: .celsius))
+        try testIdentity(from: TemperatureMeasurement(temperature: data, unit: .fahrenheit))
+
+        try testIdentity(from: TemperatureMeasurement(temperature: data, unit: .celsius, timeStamp: time, temperatureType: .ear))
+        try testIdentity(from: TemperatureMeasurement(temperature: data, unit: .celsius, temperatureType: .ear))
+        try testIdentity(from: TemperatureMeasurement(temperature: data, unit: .celsius, timeStamp: time))
+    }
+
+    func testTemperatureType() throws {
+        for type in TemperatureType.allCases {
+            try testIdentity(from: type)
+        }
+    }
+}
